# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Routes, note sets and notes for the RecceLog application.",
        "title": "RecceLog API",
        "contact": {
            "name": "RecceLog"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/users/me": {
            "get": {
                "description": "Returns the profile of the authenticated caller (the local user provisioned from the Keycloak token).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get the current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.UserResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    }
                }
            }
        },
        "/v1/users/{id}": {
            "get": {
                "description": "Returns the public profile (display name, description, avatar) of a user by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get a user's public profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid user id",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    }
                }
            }
        },
        "/v1/users/{id}/profile_pic": {
            "get": {
                "description": "Returns the user's avatar image (custom if set, otherwise the default).",
                "produces": [
                    "image/png",
                    "image/jpeg"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get a user's profile picture",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image bytes"
                    },
                    "400": {
                        "description": "Invalid user id",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_RecceLog_api_internal_http_dto.UserResponse": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "description": "AvatarURL is the canonical, always-present URL of the user's profile\npicture. It is derived from the id (not stored): the endpoint serves the\ncustom avatar when one exists, otherwise the default image.",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_RecceLog_api_internal_http_problem.Problem": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "violations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Violation"
                    }
                }
            }
        },
        "github_com_RecceLog_api_internal_http_problem.Violation": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}